+++ src/config/definitions.local.dist.php
@@ -20,5 +20,30 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/athene2 for the canonical source repository
  */
-$moduleCache = true;
-$configCache = false;
+
+$env = 'development';
+
+$assets = [
+    'bundle_host' => 'https://packages.serlo.org/athene2-assets@a/',
+    'assets_host' => 'https://assets.serlo.org/athene2-assets/',
+    'legacy_editor_renderer' => 'http://legacy-editor-renderer:3000/',
+    'editor_renderer' => 'http://editor-renderer:3000/',
+];
+
+$db = [
+    'host' => 'mysql',
+    'port' => '3306',
+    'username' => 'root',
+    'password' => 'secret',
+    'database' => 'serlo',
+];
+
+$recaptcha = [
+    'key' => '6LfwJFwUAAAAAKHhl-kjPbA6mCPjt_CrkCbn3okr',
+    'secret' => '6LfwJFwUAAAAAPVsTPLe00oAb9oUTewOUe31pXSv'
+];
+
+$smtp_options = [];
+
+// $sentry_dsn
+// $newsletter_key