+++ src/module/Markdown/config/module.config.php
@@ -25,7 +25,6 @@ namespace Markdown;
 return [
     'service_manager' => [
         'factories' => [
-            __NAMESPACE__ . '\Options\ModuleOptions'     => __NAMESPACE__ . '\Factory\ModuleOptionsFactory',
             __NAMESPACE__ . '\Storage\MarkdownStorage'   => __NAMESPACE__ . '\Factory\MarkdownStorageFactory',
             __NAMESPACE__ . '\Service\HtmlRenderService' => __NAMESPACE__ . '\Factory\HtmlRenderServiceFactory',
             __NAMESPACE__ . '\Service\OryRenderService' => __NAMESPACE__ . '\Factory\OryRenderServiceFactory',